requirements.txt
tensorflow>=2.16.1
pillow>=10.3.0
numpy>=1.26.4
matplotlib>=3.9.2
seaborn>=0.13.2
scikit-learn>=1.5.2
xgboost>=2.1.1
tqdm>=4.67.1
streamlit>=1.39.0